Update flexibility documentation to carrier perspective
In the system perspective of flexibility a technology is considered flexible across the energy system if it is flexible for one energy carrier. This creates consistency across the energy carriers. In the carrier perspective the classification of a technology can change per carrier. This makes it easier to assess flexibility needs per carrier